commit end of experimentation
trying to make the collision system from scratch lead to very glichi experience: needs to tewet a lot but in general I like the movement and collisions of the paddels.  Anyway, moving on the collisons for the boings has to be just calculate by its radio and dont calculate if has a pixel in the sprite area, also the blocs needs a more collisions points in the ball to be efective and less glichi, but alteranete those collission points bettwwen frames due to better performance.